Refactor: change to manage stations state by App
