+++ src/components/StationManager.js
@@ -1,54 +1,33 @@
 import StationNameInput from "./StationNameInput.js";
 import StationList from "./StationList.js";
-import localStorageManager from "../util/localStorage.js";
-import { STORAGE_KEY } from "../util/constants.js";
 
-export default function StationManager({ $target, isShow }) {
+export default function StationManager({ $target, stations, isShow, onAddStation, onDeleteStation }) {
   this.$container = document.createElement("div");
   this.$container.className = "station-management";
   $target.append(this.$container);
 
   this.isShow = isShow;
-  this.stations = localStorageManager.getItem({
-    key: STORAGE_KEY.station,
-    defaultValue: [],
-  });
+  this.stations = stations;
 
   this.isExistStationName = (stationName) => {
     return this.stations.includes(stationName);
   };
 
-  this.onAddStation = (stationName) => {
-    const nextStations = [...this.stations, stationName];
-    this.setState({ nextStations });
-  };
-
-  this.onDeleteStation = (stationIndex) => {
-    const nextStations = [...this.stations];
-    nextStations.splice(stationIndex, 1);
-
-    this.setState({ nextStations });
-  };
-
   this.stationNameInput = new StationNameInput({
     $target: this.$container,
     isExistStationName: this.isExistStationName,
-    onAddStation: this.onAddStation,
+    onAddStation: onAddStation,
   });
+
   this.stationList = new StationList({
     $target: this.$container,
     stations: this.stations,
-    onDeleteStation: this.onDeleteStation,
+    onDeleteStation: onDeleteStation,
   });
 
   this.setState = ({ nextIsShow, nextStations }) => {
     if (nextStations) {
       this.stations = nextStations;
-      localStorageManager.setItem({
-        key: STORAGE_KEY.station,
-        item: nextStations,
-      });
-
       this.stationList.setState({ nextStations: this.stations });
     }
